implication removed, added new rule to express variable domains, created a folder for machine_graph util classes, added new class that iteratively creates graphs with all possible combination of pumps and valves values, added new attributes to TubeEdge so cutting and changing direction of an edge is possible without changing its original value. Create util class sequence that returns all values in a sequence of numbers: Sequence(-1,1) = -1,0,1. Fixed some compiling errors. Still not tested.

+++ rules/arithmetic/integernumber.cpp
@@ -1,6 +1,6 @@
 #include "integernumber.h"
 
-IntegerNumber::IntegerNumber(int value) :
+IntegerNumber::IntegerNumber(long long value) :
     ArithmeticOperable()
 {
     this->value = value;
